requirements.txt
python-jobspy>=1.1.75
httpx>=0.27.0
anthropic>=0.40.0
click>=8.1.0
thefuzz[speedup]>=0.22.0
python-Levenshtein>=0.25.0
